Consolidate all client initialization code to a single singleton method in the base and move all setup and teardown to base as well
