+++ tests/test_dynamodb_full_table_sync.py
@@ -13,8 +13,6 @@ import pdb
 import paramiko
 import csv
 import json
-import boto3
-from boto3.dynamodb.types import TypeSerializer
 from datetime import datetime, timedelta, timezone
 from singer import utils, metadata
 import singer
@@ -34,30 +32,6 @@ class DynamoDBFullTable(TestDynamoDBBase):
             'generator': self.generate_items,
             'num_rows': 3531},
         ]
-
-    def setUp(self):
-        client = boto3.client('dynamodb',
-                              endpoint_url='http://localhost:8000',
-                              region_name='us-east-1')
-
-        table_configs = self.expected_table_config()
-
-        self.clear_tables(client)
-
-        for table in table_configs:
-            self.create_table(client,
-                              table['TableName'],
-                              table['HashKey'],
-                              table['HashType'],
-                              table.get('SortKey'),
-                              table.get('SortType'))
-
-        waiter = client.get_waiter('table_exists')
-        for table in table_configs:
-            LOGGER.info('Adding Items for {}'.format(table['TableName']))
-            waiter.wait(TableName=table['TableName'], WaiterConfig={"Delay": 1, "MaxAttempts": 20})
-            for item in table['generator'](table['num_rows']):
-                client.put_item(TableName=table['TableName'], Item=item['M'])
 
     def name(self):
         return "tap_tester_dynamodb_full_table"
@@ -189,11 +163,5 @@ class DynamoDBFullTable(TestDynamoDBBase):
             # assert that there is a version bookmark in state
             first_versions[table_name] = state['bookmarks'][table_name]['version']
             self.assertIsNotNone(first_versions[table_name])
-
-        client = boto3.client('dynamodb',
-                              endpoint_url='http://localhost:8000',
-                              region_name='us-east-1')
-
-        self.clear_tables(client)
 
 SCENARIOS.add(DynamoDBFullTable)
