+++ tests/test_dynamodb_log_based.py
@@ -13,12 +13,12 @@ import pdb
 import paramiko
 import csv
 import json
-import boto3
-from boto3.dynamodb.types import TypeSerializer
 from datetime import datetime, timedelta, timezone
 from singer import utils, metadata
 import singer
 import decimal
+
+from boto3.dynamodb.types import TypeSerializer
 
 from base import TestDynamoDBBase
 
@@ -42,79 +42,6 @@ class DynamoDBLogBased(TestDynamoDBBase):
                 'int_id': i,
             }
             yield serializer.serialize(record)
-
-    def setUp(self):
-        client = boto3.client('dynamodb',
-                              endpoint_url='http://localhost:8000',
-                              region_name='us-east-1')
-
-        table_configs = self.expected_table_config()
-
-        self.clear_tables(client)
-
-        for table in table_configs:
-            self.create_table(client,
-                              table['TableName'],
-                              table['HashKey'],
-                              table['HashType'],
-                              table.get('SortKey'),
-                              table.get('SortType'))
-
-        waiter = client.get_waiter('table_exists')
-        for table in table_configs:
-            LOGGER.info('Adding Items for {}'.format(table['TableName']))
-            waiter.wait(TableName=table['TableName'], WaiterConfig={"Delay": 1, "MaxAttempts": 20})
-            for item in table['generator'](table['num_rows']):
-                client.put_item(TableName=table['TableName'], Item=item['M'])
-
-    def addMoreData(self, numRows):
-        client = boto3.client('dynamodb',
-                              endpoint_url='http://localhost:8000',
-                              region_name='us-east-1')
-
-        table_configs = self.expected_table_config()
-
-        for table in table_configs:
-            LOGGER.info('Adding Items for {}'.format(table['TableName']))
-            for item in table['generator'](numRows, table['num_rows']):
-                client.put_item(TableName=table['TableName'], Item=item['M'])
-
-    def updateData(self, numRows):
-        client = boto3.client('dynamodb',
-                              endpoint_url='http://localhost:8000',
-                              region_name='us-east-1')
-
-        table_configs = self.expected_table_config()
-
-        for table in table_configs:
-            LOGGER.info('Adding Items for {}'.format(table['TableName']))
-            for item in table['generator'](numRows):
-                client.put_item(TableName=table['TableName'], Item=item['M'])
-
-    def deleteData(self, id_range):
-        client = boto3.client('dynamodb',
-                              endpoint_url='http://localhost:8000',
-                              region_name='us-east-1')
-
-        for table in self.expected_table_config():
-            for id in id_range:
-                client.delete_item(TableName=table['TableName'],
-                                   Key={'int_id': {
-                                       'N': str(id)}})
-
-    def disableStreams(self, table_names):
-        client = boto3.client('dynamodb',
-                              endpoint_url='http://localhost:8000',
-                              region_name='us-east-1')
-
-        for table_name in table_names:
-            client.update_table(
-                TableName=table_name,
-                StreamSpecification={
-                    'StreamEnabled': False,
-                },
-            )
-
 
     def name(self):
         return "tap_tester_dynamodb_log_based"
@@ -202,8 +129,11 @@ class DynamoDBLogBased(TestDynamoDBBase):
                                                                                    annotated_schema,
                                                                                    additional_md)
 
+        # Disable streams forces shards to close
+        self.disableStreams(expected_streams)
         # run full table sync
         sync_job_name = runner.run_sync_mode(self, conn_id)
+        self.enableStreams(expected_streams)
 
         exit_status = menagerie.get_exit_status(conn_id, sync_job_name)
         menagerie.verify_sync_exit_status(self, exit_status, sync_job_name)
@@ -250,7 +180,10 @@ class DynamoDBLogBased(TestDynamoDBBase):
         ################################
         # Run sync again and check that no records came through
         ################################
+        # Disable streams forces shards to close
+        self.disableStreams(expected_streams)
         sync_job_name = runner.run_sync_mode(self, conn_id)
+        self.enableStreams(expected_streams)
 
         exit_status = menagerie.get_exit_status(conn_id, sync_job_name)
         menagerie.verify_sync_exit_status(self, exit_status, sync_job_name)
@@ -277,6 +210,8 @@ class DynamoDBLogBased(TestDynamoDBBase):
         ################################
         # Run sync again and check that records did come through
         ################################
+        # Disable streams forces shards to close
+        self.disableStreams(expected_streams)
         sync_job_name = runner.run_sync_mode(self, conn_id)
 
         exit_status = menagerie.get_exit_status(conn_id, sync_job_name)
@@ -290,12 +225,5 @@ class DynamoDBLogBased(TestDynamoDBBase):
             self.assertEqual(31, len(stream['messages']))
 
         state = menagerie.get_state(conn_id)
-
-        # TODO Check log based things
-        client = boto3.client('dynamodb',
-                              endpoint_url='http://localhost:8000',
-                              region_name='us-east-1')
-
-        self.clear_tables(client)
 
 SCENARIOS.add(DynamoDBLogBased)
