+++ tests/test_dynamodb_projections.py
@@ -13,12 +13,12 @@ import pdb
 import paramiko
 import csv
 import json
-import boto3
-from boto3.dynamodb.types import TypeSerializer
 from datetime import datetime, timedelta, timezone
 from singer import utils, metadata
 import singer
 import decimal
+
+from boto3.dynamodb.types import TypeSerializer
 
 from base import TestDynamoDBBase
 
@@ -69,30 +69,6 @@ class DynamoDBProjections(TestDynamoDBBase):
                 'null_field': None
             }
             yield serializer.serialize(record)
-
-    def setUp(self):
-        client = boto3.client('dynamodb',
-                              endpoint_url='http://localhost:8000',
-                              region_name='us-east-1')
-
-        table_configs = self.expected_table_config()
-
-        self.clear_tables(client)
-
-        for table in table_configs:
-            self.create_table(client,
-                              table['TableName'],
-                              table['HashKey'],
-                              table['HashType'],
-                              table.get('SortKey'),
-                              table.get('SortType'))
-
-        waiter = client.get_waiter('table_exists')
-        for table in table_configs:
-            LOGGER.info('Adding Items for {}'.format(table['TableName']))
-            waiter.wait(TableName=table['TableName'], WaiterConfig={"Delay": 1, "MaxAttempts": 20})
-            for item in table['generator'](table['num_rows']):
-                client.put_item(TableName=table['TableName'], Item=item['M'])
 
     def name(self):
         return "tap_tester_dynamodb_projections"
@@ -239,11 +215,5 @@ class DynamoDBProjections(TestDynamoDBBase):
                         for list_key in config['top_level_list_keys']:
                             self.assertTrue(isinstance(message['data'][list_key], list))
                         self.assertEqual(config['nested_map_keys']['map_field'], {*message['data']['map_field'].keys()})
-
-        client = boto3.client('dynamodb',
-                              endpoint_url='http://localhost:8000',
-                              region_name='us-east-1')
-
-        self.clear_tables(client)
 
 SCENARIOS.add(DynamoDBProjections)
